https://github.com/kswoll/WootzJs/issues/16
Fix bug when generating the default constructor for a class.  Had been checking for whether or not the class had any ConstructorDeclarationSyntax members, without ensuring that they were not static.  Thus any class that had no instance constructors but *did* have a static constructor would not have the default constructor generated.  Updated TypeTests to account for this bug.